Search: integrate under `/`
- update styles
- integrate under `/`

I copied all the code from
https://github.com/readthedocs/readthedocs-sphinx-search/pull/132/.
However, there are lot of things that are not required for this stage.
I removed some of them and make the integration nicer with the client we are
creating.

These changes are related to the HTML structure and the CSS style. The
functionality is exactly the same.

Related #19