Compatibility: send `X-RTD-Hosting-Integrations-Version` header
Send an internal HTTP header with the current `package.json`'s version. This
will allow us to decide in the server what's the correct JSON structure for this
particular version of the client.

This way, we can keep support for older versions allowing us to move forward
expanding the features of the client and making breaking changes releases as
well. As long as the API response keeps the same for older versions.

Closes #28