Search: use `em` instead of `rem` unit
Using `rem` makes the value depend on an element (the root element) we don't
have control over. So, the modal was looking different in Sphinx than on
Material for MkDocs.

I migrated all the `rem` to `em` by taking a look at the calculated pixel unit
and adapted them to keep looking in the same way. `em` takes as a reference the
parent of the element, which we do have control.

https://developer.mozilla.org/en-US/docs/Learn/CSS/Building_blocks/Values_and_units