HTML attributes API: configure the addon via HTML attributes
By default the search modal is shown when pressing `/`.
However, we want this to be configurable by the user.
This can be achieved by passing `show-modal-keycode` attribute to the web
component as follows:

```html
<readthedocs-search show-modal-keycode="220">
</readthedocs-search>
```

In this case, it will be shown when pressing `\`.

I think this is a good pattern for static values that won't be changed from the
dashboard or similar in a dynamic way.

Related #13