Switch to web-test-runner test suite
This switches out Jest for web-test-runner, which is the test suite that
is recommended for web component building, especially from Lit:
https://lit.dev/docs/tools/testing/

This replaces manual DOM testing manipulation and instead tests inside
the browser environment instead. This is handy for picking up on
exceptions that wouldn't be caught outside the browser -- I picked up on
a couple just adding notification tests.

The Lit starter kit has some more examples of tests:
https://github.com/lit/lit-element-starter-js/blob/main/test/my-element_test.js

I just added some to test the functionality. This still requires some
mocking, as all imports do need to be JS modules -- CSS and JSON
included in that.