Embeded webcomponents
This is an initial attempt to be able to render the `readthedocs-search` addon
inside the `readthedocs-flyout` one.

I was able to:

- include the `readthedocs-search` tag inside the flyout
- connect the `focusin` event to the input from inside the flyout to call the search

Each time I click on the input from inside the flyout, it shows a log in the
console (for debugging). However, due to some reason that I don't know yet, the
`readthedocs-search` tag does not contains any HTML inside ito, so even when the
modal is triggered, nothing is shown.

Related #90