Remove check for "compiled assets up-to-date"
Do not check if the `dist/` compiled assets are up-to-date on CircleCI.
This has been pretty annoying to work with.
Each time you add a commit, you have to re-compile the assets.
Then if another PR is merged, you have to update your branch and re-compile the
assets.

Now, since we are adding a release process that explicity re-compile the assets
before doing the release, compiling them on each PR doesn't make sense anymore.

We will be always using the compiled assets from GitHub tags now when deploying.