Swap the flyout version to be a `code-branch` icon
This will look cleaner and more obvious to users than the `v:` that it means a version.
I couldn't get the actual fontawesome icon to load,
but wanted to open this PR as a starting point.

If we're not loading Fontawesome icons via CDN,
which makes sense for performance,
we can perhaps load it via svg?