Notification: simple stable-latest version warning logic
I implemented the simple idea exposed in #71.
I think we can start with this simple logic that's easy to explain and grow from
here if necessary.

- on `latest`: a notification to warn the user about some features may not yet
  be available is shown
- on non-`stable` nor `latest` versions: a notification to warn the user about
  possible reading and old/outdated version is shown

Note that both notification are pretty simple, it does not require external
dependencies (e.g. `semver`) and it's easy to explain to users. The warning is
shown only if `stable` is an active version --since we link to it from both.

_However_, I think the best way for users to make usage of this
addon is by implementing the logic that fits their needs by themselves.
We don't allow users to customize this logic just yet, but we plan to do it in
the future when the addons is more settled down.

Closes #71