EthicalAds: follow the "exception handler" pattern we disccuss
The idea here is to check for all the _mandatory_ fields inside `.isEnabled()`
method and return `false` if any of the _mandatory_ fields does not exist.

`objectPath` handles non-existent nested fields and returns the default
value of any of the nested fields don't exist.

There are other attributes like `keywords` and `campain_type` that are
_optional_, so they are checked at render time by the addon itself.

Related #122