Move HTML attributes outside the constructor
The DOM does not support to add HTML elements with attributes.
They have to be enabled without attributes and then those attributes have to be
added.

Otherwise, the following exception is raised:

```
DOMException: Failed to construct 'CustomElement': The result must not  have
attributes
```

This commit moves the `className` attribute to `loadConfig` which is executed
_after_ the element is injected in the DOM.