Fix a few issues with our prettier and pre-commit configuration (#156)
I found a few issues trying to bring these additions into the new
dashboard
today:

- Pre-commit is configured to point to a specific ref instead of using
the `local` target. The `local` target seems better for files in the
same repository as the pre-commit config. This allows the hooks file
to be dropped. The script can be moved into a hidden path here as
well.
- Prettier wasn't configured correctly and seemed to be missing files.
Instead of passing paths as arguments, reuse the configuration from
the new dashboard and rely on .prettierignore to remove files from
Prettier formatting.
- Rerun prettier with all found files