API: send `project-slug=` and `version-slug=` to the backend
Follow the plan discussed in
https://github.com/readthedocs/readthedocs.org/issues/10536

Grab the `project-slug` and `version-slug` from the `meta` tag injected by CF
and send it back to the backend. If there is any addons that can be enabled in
the current page that depends on the URL, it will send it also.

Each Addon class should implement `requireUrlParam` to decide whether or not to
send the `url=` to the backend API or not.