JSON schema and validation
Define a JSON schema and use `ajv` to load/compile it.
Then, use this schema to validate the response returned by the API.
If the response does not validate the schema we throw an exception and don't
load any of the addons.

I plan to modify this to validate each part of the response by the addon itself
and disable a particular addon if the validation fails. This will allow having a
broken addon without interfering with the rest.

Related #149