EthicalAds: support explicit placement
Use the `#ethical-ad-placement` explicit placement on Read the Docs to keep
backward compatibility with current integrations.

Users just need to define a `div` whenever they want in the HTML and style with
the CSS classes they want as follows:

```
<div id="ethical-ad-placement" class="ad-flat"></div>
```

This will tell our EthicalAdsAddon to load the ad in that position.
Note that `data-ea-manual="true"` should not be used on these cases.

Closes https://github.com/readthedocs/meta/issues/132