Test: analytics with mocked requests
After researching a lot I found the pattern that I like the most.

* it uses `sinon`, which is a standard library/framework for mocking
* it uses `sinon.fakeServer` to mock requests/responses easily, giving us a lot
  of extra features to inspect the mocked URL and assert multiple fields
* since `sinon` doesn't support (yet) mocking global `fetch`,
  I'm using `unfetch`, which is a small library that converts `fetch` into
  `XMLHttpRequest` call by imitating the same API than global `fetch`
* once `sinon` add support for this, we just need to remove an import

I'm happy with this pattern and I will keep expanding it as we go. Nothing is
written in stone here, but I've tested all the other different patterns we
talked about and they are are dirty to me, so I wanted to keep it simple here
for now while having the `sinon.fakeServer` features that I like.