Notifications: respect `project.versioning_scheme` for links
This is a workaround for now since the backend API is not returning this URLs
due to https://github.com/readthedocs/readthedocs-ops/issues/1323. However, we
cannot fix that quickly.

I'm re-implementing a small version of the resolver for this by generating the
URL based on the `project.versioning_scheme` for now.

We should come back to this once the API returns the URLs we need.

Closes #243