EthicalAds: inject ad only if it's a known theme and above the fold
Disable using `stickybox` style for now and only inject the ad if we are sure
it's a known theme and the ad it's going to be above the fold.

In a following iteration, we will want to use `stickybox` in the other cases but
we want to have the ability to disable the re-placement in smaller screens.

Closes #328