Fix recent searches when looking at different versions on same domain (#353)
Fixes #296 

A few considerations here:
- I used the new localStorage code that we added in #283 . This means
the key for localStorage will change once this is deployed, meaning that
previous recent searches won't show up anymore (and there will be an
unused object in localStorage lying around). Are we ok with that, or do
we want to make it backward compatible?
- I didn't add any logic for removing recent searches from previous
versions. If we don't add anything, this object will grow and grow and
it will store recent searches of versions that might become very old. Do
we want to add some kind of logic in this PR or should we do it in a
separate PR?
- Besides project and version, I also added language to the key (the
same way we do it for notifications). Is that ok?

---------

Co-authored-by: Manuel Kaufmann <humitos@gmail.com>