Notification: auto-dismiss while visible
It detects if the tab is opened and trigger a timeout for 5 seconds. After that
time, the notification is hidden.

Eventually, we could use some nice effect, like fade out and stop the timer
while the pointer is over the notification -- but that't outside the scope of
this initial work.

Closes #366