Links to new dashboard
Now that we are using El Proxito with the new templates, we end up with
`app.app.readthedocs.org` links.

This is an improvement of that simple replacement to remove the duplicated
`app.` from the resulting link.

I want to keep this code around as a way to keep promoting the new dashboard.
Once we are fully on the new templates, we can remove this code.

Closes #380