Notification: don't show it on `Project.default_version`
Don't show the notification "This may be an old version of this documentation"
if the reader is currently reading the `Project.default_version`.

I'm on the fence about doing this part of the default logic or a config on its
own. I'll probably change it to a config after thinking a little more, but
pushing these changes for now.

Closes #406