Fix the filters title covering other labels (#354)
Fix #293 

There's still a slight difference. The only way to avoid that
completely, would be to remove the default checkbox, and style our own.
But I'm not really keen on that solution.

BTW this styling/positioning will probably be further improved in #292 ,
so this is just a quick solution to make it look better.