File Tree Diff: initial implementation (#424)
Added a new addon to show all the filenames and URLs that were added,
deleted or modified in the current PR compared to the LATEST version.


![Screenshot_2024-11-07_12-44-02](https://github.com/user-attachments/assets/dfdba906-1eb1-4755-a1e6-eaddbfa51144)

I'm rendering this data in a new notification decoupled from the
existing one on purpose. I don't want to entangle the code of these two
different addons for now, in particular while we are under
development/testing. I also see that in the future they may have
different UI elements, or the data will be integrated in a more bigger
UI element (eg. toolbar) or similar.

We can discuss more how this addon will grow and where is a good place
to add this data and make that work in a following PR. For now, I want
something that we can start testing internally in our own projects.

### Examples


![Screenshot_2024-11-07_14-15-37](https://github.com/user-attachments/assets/d0265b34-8ade-4348-bc8e-e6a3ff8ae0d1)



![Screenshot_2024-11-07_14-35-41](https://github.com/user-attachments/assets/403b6860-60c1-468e-ae5d-2491f6f47b38)



![Screenshot_2024-11-07_14-51-06](https://github.com/user-attachments/assets/0e24664d-672f-40c0-9be4-bd76b2cf4f04)


Closes https://github.com/readthedocs/addons/issues/409
Requires https://github.com/readthedocs/readthedocs.org/pull/11749