Add initial Command Pallette addon
This is one I'm pretty excited about.
It takes the Command Pallette pattern from Slack, GitHub, etc.,
and adds it to any documentation page.

I think this could be an interesting UX for our future UX,
and enable a lot of really interesting power user features.

The really nice thing is that it's totally based on a keyboard shortcut,
so doesn't require any UX, and generally folks won't want to disable it.