Visual diff: trigger `DOM_CHANGED` event after enable/disable
A new event `READTHEDOCS_ROOT_DOM_CHANGED` is triggered when the "Visual diff"
is enabled/disabled after the DOM was modified.

This allows other addons to subscribe to this event to re-initialize if
required. I used this event from links preview to call `setupTooltips` again to
re-install tooltips on these links.

I migrated links preview to a `LitElement` to be able to make usage of this
pattern and follow what we already had.

We will be able to do something similar on #157