API: save a "flag" on `window` object
We need to find a way for people subscribing to the Read the Docs data ready
event _after_ it was triggered. This is an idea about how to solve this use
case.

It adds a "flag" in the `window` object that users can check _before_
subscribing to that event to make usage of the data if the event was triggered
already.

* Reference: https://github.com/readthedocs/readthedocs.org/issues/10648#issuecomment-2542003070
* Example: https://github.com/readthedocs/test-builds/blob/full-feature/docs/static/readthedocs.js