Don't show the `latest` notification if it's the default version (#480)
I think this is maybe a better default than
https://github.com/readthedocs/readthedocs.org/pull/11874.

This should at least stop the "getting a notification when I hit the
default version" issue. Though maybe it's better to default it to off,
instead of this? I think we need to do _something_ to fix the current
behavior because it's properly annoying (I hit the latest version
warning like 3 times yesterday trying to do basic work, and it was never
useful..)