Keep a copy of addons instances and re-load them
The `AddonsApplication` keeps a list of all the addons instances created when
the page was loaded (`AddonsApplication.addonsInstances`).

Then, when the `READTHEDOCS_URL_CHANGED` event is fired, we check if we already
have instances created and, in that case, we call `addon.loadConfig(config)` to
re-configure those instances with the new `config` object --without creating new
ones.

This commit includes some refactor as well:

- Share all the `constructor()` method all accross addons
- Override the `constructor()` for those addons that are not LitElements
- Create `loadConfig()` methods for those addons that are not LitElements
- Remove code commented out