File Tree Diff: simplify and fix the algorithm to find chunks (#527)
- We get all the elements with `.doc-diff-*` classes
- For each element, we find the immediate parent matching a "section
node"
- De-duplicate those nodes (multiple changes in the same section)

This approach doesn't require creating a big CSS query selector with
multiple `:has(.doc-diff-*)` pseudo-selector. I think it's a lot easier
to understand the logic behind it than what I wrote originally. Besides,
it solves a few false positives we had in the previous approach where we
were detecting the same change twice but on different sections like
`section` _and_ `h2` for the same change.

This is still not perfect, but it's better than what we currently have.
There is going to be more iterations on this since the behavior in not
100% defined yet, but we are getting closer.

## Examples that are fixed with this PR

Instead of showing 4 chunks here, it shows only the 2 inner `p`.


https://cpython-previews--129935.org.readthedocs.build/en/129935/c-api/veryhigh.html?readthedocs-diff=true&readthedocs-diff-chunk=1

### Old behavior

[Peek 2025-02-10
16-29.webm](https://github.com/user-attachments/assets/fc071de8-475c-4560-add9-3b99b2729696)

### New behavior

[Peek 2025-02-10
16-32.webm](https://github.com/user-attachments/assets/e2b9a3f5-9d16-4ebf-890e-afd9d17fed53)

---------

Co-authored-by: Eric Holscher <25510+ericholscher@users.noreply.github.com>