Resolver filename fixes (#532)
Make flyout work when `readthedocs.resolver.filename` is `null`. In that
case, we get this value from the META tag added by Cloudflare.

This was the original intent here, but I missed allowing `null` in the
data validation. Besides, even if the data validation allows `null`,
there was another bug since I was re-defining the variable (`const ...`)
instead of overriding its value.

We didn't catch this locally because we have a different rule that makes
the request to always contain `url=` parameter and that case is not
affected by this.

Besides, I didn't find this when doing QA after deploy because I tested
with examples that sent `url=` parameter.

I'll see if I can write a test for this in another PR.