Always add the tooltips before closing `body`
There is no need to add the tooltip next to the anchor element where the tooltip
is going to be displayed. The position is pre-calculated and it does not
interfere where in the DOM is added.

To simplify, I'm adding the tooltips always before closing the body of the page.