VisualDiff: complex selector and `?maincontent=` API argument (#596)
When we send the `?maincontent=` CSS selector to the API endpoint, the
response file is chunked to only return the chunk that contains the CSS
selector.

There are cases when we use inner CSS selector that won't work on the
returned response (eg. `main > div > div.md-content`) since the response
will contain only the `<div class="md-content">...` part.

In those cases, the query for `this.rootSelector` will fail and we will
fallback to the `document.body.firstElementChild` which will always be
the element we are looking for.

Unfortunately, I wasn't able to write a test case for this. I've tried
different approaches and wasn't able to :/

Closes #576