Reintroduce the stickybox on wide screens (#604)
Alternates between using the stickybox placement and the footer
placement on unknown themes. It also adds an ID to the stickybox
placement so we can see performance of it.

As written, this introduces a random element which may not be desirable.
We could instead introduce the stickybox only on very wide screens (say
>=1600px) or we could base it on the title of the project or something
like that to make it deterministic. I'm open to suggestions.


>         // TODO: Check this placement on the dashboard,
>         // and see how this is performing.

The footer placement is performing very well at ~4x the CTR with ~50%
the view rate vs. previously.

Ref: https://github.com/readthedocs/addons/pull/562