Search: filters enabled by default (#612)
Read the addons API and enable/disable the filters by default based on
that data. This is an initial and immediate simple solution for #22
without implementing all the features.

Basically, all projects that have subprojects will now have "Include
subprojects" checked by default. In the future, we will allow users to
enable/disable this per project.